rename junit4 module
