+++ .github/workflows/junit4.yml
@@ -16,7 +16,7 @@ jobs:
       - uses: axel-op/googlejavaformat-action@v3
         with:
           args: "--replace"
-          files: "selenium-junit4-examples/**/*.java"
+          files: "selenium-junit4/**/*.java"
           skip-commit: true
       - name: Print diffs
         run: git --no-pager diff --exit-code
@@ -30,7 +30,7 @@ jobs:
         with:
           java-version: 11
       - name: Run JUnit 4 Tests
-        working-directory: ./selenium-junit4-examples
+        working-directory: ./selenium-junit4
         env:
           SAUCE_USERNAME: ${{ secrets.SAUCE_USERNAME }}
           SAUCE_ACCESS_KEY: ${{ secrets.SAUCE_ACCESS_KEY }}
