+++ selenium-junit4/src/test/java/com/saucedemo/selenium/junit4/CheckoutTest.java
@@ -1,4 +1,4 @@
-package com.saucedemo.selenium.junit4.demo;
+package com.saucedemo.selenium.junit4;
 
 import org.junit.Assert;
 import org.junit.Before;
